requirements.txt
chromadb>=0.4.22
sentence-transformers>=2.3.1
pyyaml>=6.0.1
python-dotenv>=1.0.0
openai>=1.12.0
langchain>=0.1.0
langchain-community>=0.0.20
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
numpy>=1.24.0
tqdm>=4.66.0

